+++ src/components/CartIconWithCount/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './CartIconWithCount';
